C++: Turn header variant tests that use PCH files into integration tests
These tests currently test a code path in the extractor that only exists for
these tests. By turning them into integration tests, we actually use the
code path that normal database creation uses.
