+++ cpp/ql/integration-tests/header-variant-tests/clang-pch/a.c
@@ -1,0 +1,2 @@
+#include "a.h"
+#define FOUR 4
