+++ cpp/ql/integration-tests/header-variant-tests/clang-pch/a.h
