+++ cpp/ql/integration-tests/header-variant-tests/clang-pch/c.c
@@ -1,0 +1,3 @@
+int main() {
+  return ONE + FOUR;
+}
