+++ cpp/ql/integration-tests/header-variant-tests/clang-pch/clang-pch.expected
