+++ cpp/ql/integration-tests/header-variant-tests/clang-pch/d.c
@@ -1,0 +1,1 @@
+#import "d.h"
