+++ cpp/ql/integration-tests/header-variant-tests/clang-pch/e.c
@@ -1,0 +1,3 @@
+int main() {
+  return SEVENTEEN;
+}
