+++ cpp/ql/integration-tests/header-variant-tests/clang-pch/f.c
@@ -1,0 +1,5 @@
+#if 1
+#pragma hdrstop
+extern int x;
+#define SEEN_F
+#endif
