+++ cpp/ql/integration-tests/header-variant-tests/clang-pch/g.c
@@ -1,0 +1,5 @@
+#ifdef SEEN_F
+static int g() {
+  return 20;
+}
+#endif
