+++ cpp/ql/integration-tests/header-variant-tests/clang-pch/h.c
@@ -1,0 +1,4 @@
+#include "h1.h"
+#pragma hdrstop
+#include "h2.h"
+#define SEEN_H
