+++ cpp/ql/integration-tests/header-variant-tests/clang-pch/h2.h
