+++ cpp/ql/integration-tests/header-variant-tests/microsoft-pch/a.c
@@ -1,0 +1,1 @@
+#include "a.h"
