+++ cpp/ql/integration-tests/header-variant-tests/microsoft-pch/a.h
