+++ cpp/ql/integration-tests/header-variant-tests/microsoft-pch/b.c
@@ -1,0 +1,6 @@
+#pragma hdrstop
+#include "b.h"
+
+int b() {
+  return A;
+}
