+++ cpp/ql/integration-tests/header-variant-tests/microsoft-pch/c.c
@@ -1,0 +1,6 @@
+#include "d.h"
+#include "c.h"
+
+int c() {
+  return A;
+}
