+++ cpp/ql/integration-tests/header-variant-tests/microsoft-pch/microsoft-pch.expected
