--- cpp/ql/test/header-variant-tests/clang-pch/_.c
@@ -1,3 +1,0 @@
-// This file exists to ensure that the output subdirectory exists prior to
-// a.c being indexed, as said directory needs to exist for the PCH file to
-// be created, and will be created by running the extractor.
