--- cpp/ql/test/header-variant-tests/clang-pch/a.c
@@ -1,3 +1,0 @@
-#include "a.h"
-#define FOUR 4
-// semmle-extractor-options: --clang -emit-pch -o ${testdir}/clang-pch.testproj/a.pch
