--- cpp/ql/test/header-variant-tests/clang-pch/c.c
@@ -1,4 +1,0 @@
-int main() {
-  return ONE + FOUR;
-}
-// semmle-extractor-options: --clang -include ${testdir}/clang-pch.testproj/a -Iextra_dummy_path
