--- cpp/ql/test/header-variant-tests/clang-pch/d.c
@@ -1,2 +1,0 @@
-#import "d.h"
-// semmle-extractor-options: --clang -emit-pch -o ${testdir}/clang-pch.testproj/d.pch
