--- cpp/ql/test/header-variant-tests/clang-pch/e.c
@@ -1,4 +1,0 @@
-int main() {
-  return SEVENTEEN;
-}
-// semmle-extractor-options: --clang -include-pch ${testdir}/clang-pch.testproj/d.pch
