--- cpp/ql/test/header-variant-tests/clang-pch/g.c
@@ -1,6 +1,0 @@
-#ifdef SEEN_F
-static int g() {
-  return 20;
-}
-#endif
-// semmle-extractor-options: --clang -include-pch ${testdir}/clang-pch.testproj/f.pch
