--- cpp/ql/test/header-variant-tests/clang-pch/h.c
@@ -1,5 +1,0 @@
-#include "h1.h"
-#pragma hdrstop
-#include "h2.h"
-#define SEEN_H
-// semmle-extractor-options: --clang -emit-pch -o ${testdir}/clang-pch.testproj/h.pch
