--- cpp/ql/test/header-variant-tests/microsoft-pch/a.c
@@ -1,2 +1,0 @@
-#include "a.h"
-// semmle-extractor-options: --microsoft /Yca.h /Fp${testdir}/microsoft-pch.testproj/a.pch
