--- cpp/ql/test/header-variant-tests/microsoft-pch/b.c
@@ -1,7 +1,0 @@
-#pragma hdrstop
-#include "b.h"
-
-int b() {
-  return A;
-}
-// semmle-extractor-options: --microsoft /Yub.h /Fp${testdir}/microsoft-pch.testproj/a.pch
