--- cpp/ql/test/header-variant-tests/microsoft-pch/c.c
@@ -1,7 +1,0 @@
-#include "d.h"
-#include "c.h"
-
-int c() {
-  return A;
-}
-// semmle-extractor-options: --microsoft /Yuc.h /Fp${testdir}/microsoft-pch.testproj/a.pch
